+++ crates/bitwarden/src/auth/login/api_key.rs
@@ -1,12 +1,15 @@
+use std::str::FromStr;
+
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
-use std::str::FromStr;
 
 use crate::{
     auth::{
         api::{request::ApiTokenRequest, response::IdentityTokenResponse},
-        login::determine_password_hash,
-        response::ApiKeyLoginResponse,
+        login::{
+            determine_password_hash, response::two_factor::TwoFactorProviders,
+            PasswordLoginResponse,
+        },
     },
     client::LoginMethod,
     crypto::CipherString,
@@ -74,4 +77,28 @@ pub struct ApiKeyLoginRequest {
 
     /// Bitwarden account master password
     pub password: String,
+}
+
+#[derive(Serialize, Deserialize, Debug, JsonSchema)]
+#[serde(rename_all = "camelCase", deny_unknown_fields)]
+pub struct ApiKeyLoginResponse {
+    pub authenticated: bool,
+    /// TODO: What does this do?
+    pub reset_master_password: bool,
+    /// Whether or not the user is required to update their master password
+    pub force_password_reset: bool,
+    two_factor: Option<TwoFactorProviders>,
+}
+
+impl ApiKeyLoginResponse {
+    pub(crate) fn process_response(response: IdentityTokenResponse) -> Result<ApiKeyLoginResponse> {
+        let password_response = PasswordLoginResponse::process_response(response)?;
+
+        Ok(ApiKeyLoginResponse {
+            authenticated: password_response.authenticated,
+            reset_master_password: password_response.reset_master_password,
+            force_password_reset: password_response.force_password_reset,
+            two_factor: password_response.two_factor,
+        })
+    }
 }
