+++ crates/bitwarden/src/auth/login/response/captcha_response.rs
