+++ crates/bitwarden/src/auth/login/response/two_factor/authenticator.rs
