+++ crates/bitwarden/src/auth/login/response/two_factor/web_authn.rs
