+++ crates/bitwarden/src/auth/login/response/two_factor/yubi_key.rs
