+++ crates/bitwarden/src/auth/mod.rs
@@ -1,4 +1,3 @@
 pub(super) mod api;
 pub mod login;
-pub mod response;
 pub mod renew;
