--- crates/bitwarden/src/auth/response/mod.rs
@@ -1,7 +1,0 @@
-mod api_key_login_response;
-mod captcha_response;
-mod password_login_response;
-pub mod two_factor_login_response;
-
-pub use api_key_login_response::*;
-pub use password_login_response::*;
